Added a few specs for Object#to_yaml
Added a test yaml file and a spec to verify that YAML#parse_file returns a YAML::Syck::Map object

Updatd spec to remove before :each for now

Fixing a mistake I made when I changed the path for the test file. Did not account that another spec was relying on that.

Began adding a few specs for YAML.dump_stream

Updated use of temp file, and fixed a typo in a specification for to_yaml

Signed-off-by: Brian Ford <bford@engineyard.com>