Renamed default config files.
The default config files used to be 1.8.mspec and 1.9.mspec. MSpec
would also try to load default.mspec if it existed. This was causing
conflicts as the settings in 1.8.mspec, for example, would override
those in the implementation-specific default.mspec file. Now, the
default file to load is named as follows:

  "#{RUBY_ENGINE}.#{RUBY_VERSION.split('.')[0,2].join('.')}.mspec"