Fix yet another race condition in a threading spec:
ThreadGroup's list_spec was making the assumption that immediately after
the two threads posted to the Channel, they would then be asleep. The
spec ran through its checks and then tried to #run and #join those
threads. But since there was no guarantee either thread had actually
reached the #sleep call, there was a real chance #run would be called too
early and #join would end up waiting for a sleeping thread. The fix was
simply to wait for both threads to either be sleeping or dead, and then
proceed to the #run and #wait calls.