Adding not_compliant_on :jruby to some block arg order-of-eval specs.
I have not gotten an official ruling from ruby-core on these specs,
but Ruby 1.9 matches JRuby's behavior, which seem to be indisputably
the correct behavior. The Ruby 1.8 behavior seems completely wrong.
I'm not sure if this should be marked as "compliant_on" for whatever
"MRI" is represented as, but it's pretty hard to argue that this is
correct and not just a peculiar impl-specific behavior for MRI.