Added tags for 1.9. See below.
The 'critical' tags are specs that either hang or cause a segfault.
These can be listed with 'mspec tag --list critical'. The rest of
the tags are from a variety of issues. These are not necessarily
bugs in 1.9. Most are probably specs that need to be fixed to be
compatible with both 1.8 and 1.9. The tags do not block syntax
errors since those occur while the file is loading.