Add .uniq to some glob specs to filter out dupes.
I sent email to ruby-core about glob sometimes returning dupes. It seems
odd that there would be dupes, but MRI does appear to return them for at
least one case. If ruby-core decides glob should never return dupes, the
spec expecting dupes is a ruby_bug. If they decide it may return dupes,
the spec I was working on for JRuby is in error and uniq is correct. We
shall see.