Fixed specs passing until a thread is asleep.
According to ri Thread#status, the following values may be returned:

  "run"
  "aborting"
  "sleep"
  false
  nil

where false is returned if a thread exits normally and nil if a thread
exits with an exception. Thread#alive? returns true if a thread's
status is "run" or "sleep".

In the specs, to pass until a thread is asleep, the following is required:

  Thread.pass while t.status and t.status != 'sleep'

If the thread exits normally or with exception before this code runs,
it will not loop infinitely like it would if the following is used:

  Thread.pass until t.status == 'sleep'

Note that the following is not a substitute for the proper way above:

  Thread.pass until t.status == 'sleep' || !t.alive?

since t.alive? would return false for an aborting thread.