Change ruby_version_is to ruby_bug for SortedSet#add, #initialize.
A couple points about this commit:

1) Until version 1.9 stabilizes, we should anly be using 'rbuy_version_is
"1.9"', not "1.9.2", etc. The reason is that we will end up with too many
ruby_version_is changes that are not relevant to any implementation in
wide use. Along the same lines, the head version of 1.9 should be used
for updating the specs to 1.9.

2) The guards were change to ruby_bug because it really makes no sense to
create a SortedSet whose elements are not mutually comparable. The fact this
wasn't found/addressed until 1.9 does not really change anything in the way
we treat MRI bugs.