Hash#to_s attempts to alias Hash#inspect on 1.9.
This is a particularly confusing commit. The principle is that as of MRI 1.9
Hash#to_s is an alias for Hash#inspect. This isn't completely true, however,
due to a bug that's been reported and guarded for. On MRI 1.8 the behaviour of
Hash#inspect is similar, although it doesn't exhibit 1.9's bug, but Hash#to_s
is quite different...