Add specs for the default included and extended being private.
I know we normally don't spec these things, but in this case it
tends to be important; there are libraries that expect to do a
diff of two lists of methods, and if these things show up they
can cause bugs. If the methods are present, it can also make it
difficult to find other bugs, like methods creeping into Object
or Module, where they should not. This was the case for a JRuby
bug today, and getting the default sets of methods to match is
something I'd like to see specified to help avoid this problem
in the future.