String#unpack: Fix for 1.9's Array#to_s change.
This test previously examined a result array from #unpack that was equal to
[-Infinity]. Infinity is actually a Float object, but Ruby won't enable its
explicit creation, so the test compared the array #to_s to a stringified
representation of '-Infinity'. 1.9. changed the way arrays were stringified,
so the test broke. This fix compares the literal -Infinity returned from
String#unpack with the result of -(Float::MAX + Float::MAX), which returns the
literal -Infinity. Easy, huh?