Hash#[]=: Tweak to workaround mspec oddity
The example included an expectation that new_hash[x].should == nil . On 1.9
that didn't cause #hash to be called on the mock (x). new_hash[x] = nil works
fine. Given that the example concerned the calling of #hash rather than the
key's value, I believe this change is a reasonable workaround.