IO#read: Move @io.close to `after :every` from `after :each`
This is a particularly unsatisfactory commit. Beforehand, @io.close was being
called after every example. This raised an Errno::EBADF after the third
example when run as part of the whole IO suite on 1.9. i.e. `mspec core/io`
caused the exception, whereas `mspec core/io/read_spec.rb` didn't. I've spent
considerable time trying to debug this with no luck. I haven't even been able
to reduce it to a test case so I can file it as a bug. Evan suggests that the
next stage is to use a debugger, but I doubt I'd be successful. :-(

I believe that this change is safe because @io is instantiated anew before
every example. That no tests fail as a result of this change reinforces the
argument. That doesn't make it OK, of course, but I'm unsure what else to do
ATM.