Quarantine some specs failing for MRI on OS X
There are a couple Dir.chdir specs and File.path specs that do not
pass on OS X because the /var path is a symlink to /private/var. I
am not sure how to fix these, since expand_path does not appear to
expand these paths completely, so we will quarantine them for now.