Add a system 'sync' after a popen spec that writes to tmp files.
The spec in question does a popen('cat > somefile'), passing a block,
and then immediately after the block has terminated it tries to read
that file. I believe the spec is failing intermittently on JRuby
because the write to 'somefile' has not yet been synced to disk on
some systems. Adding system 'sync' after the popen call allows the
spec to run consistently.