Reworked fixture directories for Dir specs.
The directory fixtures for the Dir specs are much more complex
than typical. It would be possible to create the specific files
and directories needed by the individual specs, but this is
problematic because of the number of directories and files. At
the same time, the fixtures cannot simply be static files and
directories because some paths are not valid on some platforms.

The directory fixtures are temporary files and should be cleaned
up by the specs. The specs need to run individually as well as
in groups.

The method used here is to explicitly create and remove the
fixtures by using 'before :all' and 'after :all' actions. This
is explicit, whereas the previous method would "magically"
create the fixture tree when the core/dir/fixtures/common file
was loaded, but then the temporary files would be left around.