Fixed File#path specs.
This change makes the spec agnostic about whether ad hoc details
(eg was a extraneous /dir/../dir/foo used in the name) are visible.
The core behavior of the method is returning the path to the file.
There are an infinite set of paths that reference any given file
and code should not break if dir/foo is used instead of dir/../dir/foo.