Example: simplify complex specs.
Usually I am not a fan of specs like this one that loop over
data in a single example. But there are situations like this
that would be tedious if broken into multiple examples. However,
sometimes it is best to break into multiple examples and add
a special matcher to make the specs simpler.

In this case, the file handling is moved into before/after
actions, but the file contents are written for each item in
the data being iterated. A separate describe block is wrapped
in the version guard because there are likely many more 1.9-
specific (or encoding-specific) specs that need to be added.