Fix Encoding.aliases['locale'].
Encoding.aliases['locale'] returns canonical name of the locale
encoding. The locale encoding is usually decided by
Encoding.locale_charmap, but it is raw data so their values may
be different. So when compare them, we must compare their canonical
name or Encoding object.

Locale charmap is the same as Unix command `locale charmap`,
and on Windows it is CPXXX.