Fix Hash#compare_by_identity spec to not be twitchy
The spec was extremely twitchy because it actually depended on obj and
obj.dup to have hash values (and thus object_ids) that would hash to the
same bucket so that obj would get #equal? called on it. Because
object_id's are effectly random, this spec would only actually test the
thing it was testing randomly.

Using obj on the lookup guarantee's that we'll try to call #equal? on
obj.