Mark spec as unsupported on Rubinius 1.8 mode.
Since Rubinius uses Oniguruma in 1.8 and 1.9 mode and Oniguruma does not error
with this syntax, we would need to preprocess the Regexp source string to raise
this error. Since it is a behavior that would not function in MRI, not supporting
the error itself is not preventing any code from running in Rubinius that would
otherwise run in MRI.

We do have to preprocess the Regexp source string in 1.9 mode to handle things
like UTF-8 escaped codepoints and sequences. However, it doesn't seem reasonable
to add the runtime cost in 1.8 mode just to raise this error.