Fix more race issues in Process#kill specs
Make sure we only write the pid file after we've setup the signal
handler. This makes sure we're not sending a signal before the handler
is actually setup.

The other issue is that we won't return from the constructor for
Signalizer before the thread we start has properly started the
executable. Otherwise we might be trying to kill something before we're
done.