Fix yet another race condition in Process.kill specs
This time it gets even trickier. The problem is that writing the pid
file consists of two steps, first creating the file and then writing the
pid into it. POSIX file system semantics guarantee that each of those
separate operations is atomic, but that still allows for the case where
the file is created, but the content isn't written to it yet and
therefore another thread sees an empty file.

This means that when it seems an empty file, it reads the empty string,
runs .to_i on it and sets the pid to 0. This means it will kill the
process itself, causing the SIGTERM issues.

The reason that this is happening more on Travis is probably because on
Travis build machines they still use good old spinning magnetic disks.
These are much slower, greatly increasing the changes of having this
race occur. I actually managed to trigger it much more often on an old
iMac which still has a spinning disk too.

So the change for now is that we keep trying to read the file until we
actually get contents back instead of continuing on an empty file read.