Setup locking around Process.kill process
This fixes a potential race that could occur in this subprocess. The
problem is that when two signals were received at the same time, there
was a race condition.

What could happen is that when two signals were received, it would run
the signal handler twice. One handler could set signaled to true, the
other one would see signaled as false and exit before the other handler
would print the "signaled" string.

With a Mutex we can use try_lock to attempt to lock once and then exit.
This makes sure we only print "signaled" once and immediately exit.

Why not use Mutex.synchronize {} one might ask here. Well, that doesn't
work since signals are handled in the same thread, so what happens then
is that you can recursive locking errors since the Mutex is locked twice
in the same thread so the deadlock detection kicks in and causes an
exception.