Avoid handling signal twice in Process.kill specs.
The Mutex added in 57187385 fixes the race condition but loses the ability to
ensure that the signal handler only runs once. Once the mutex unlocks, the
handler code can interrupt the code implementing Process.exit! and print
'signaled' again.  We first lock the mutex, then check the flag. This ensures
there is no race and the correct value of the flag is always seen. The flag
itself ensures the handler only runs once.