Another race in Process.kill specs.
It appears that the Processs.exit! was being handled before the write
to STDOUT had finished so the specs were failing on 1.9.3 with an empty
string instead of 'signaled'. This change allows the signal handling
code to finish and the process will naturally exit the busy loop once
it has been signaled.