Array#sample & #choice spec names should describe actual spec behavior
Our test can only check that the returned value is a member of the
Array, not that it is random, so let spec title reflect that.

Likewise, using `10.times { ... }` suggests that randomness or
consistency of the return value is somehow being tested. Just checking
for membership once should be adequate, and is clearer about what it is
actually doing.