File.open on a FIFO blocks the interpeter on 1.8.7
On 1.9.3 you can run `Thread.new { File.open('some_fifo', 'w') }` and
the interpreter will return the thread object as expected, however on
1.8.7 (and, by extension ree) we discover that `Thread.new` never
returns.

I'm not exactly sure the best way to deal with this, so if there is a
better way that just bypassing this spec on 1.8 then please let me know.