Added spec: Check for not equal
Previously I wrote a spec that tests, whether the element passed to the block is
== to the first argument given to chunk.
Now I added a spec that checks, whether the element passed to the block is not
equal to the first argument given to chunk.
I think this is a bit simpler than just testing whether dup was called on the
element.