Use a more esoteric encoding in IO.new tests
This is to fix some specs that were failing on a system where the
default external encoding was already set to ISO-8859-1.

As the :internal_encoding option is ignored when it is the same as the
external encoding, certain tests would fail due to the IO object's
internal encoding being nil.