HOME is not guaranteed to not end with a slash.
These two specs assumed that the HOME variable would never end with a slash,
which is not guaranteed at all. I bumped into this while running the specs on
Amazon CentOS, though I'm not sure if it's an OS thing or because the code was
running on an external network disk (/mnt).