Removed Kernel#caller specs.
Every implementation uses significantly different methods to implement calling
methods. In MRI, numerous actual frames are invisible because they are in C,
while in Rubinius, these frames are in Ruby and are visible.

It's not realistic or useful to attempt to specify the results of caller.
Further, it's a hack in the first place and a proper object-oriented API for
querying the stack should be created.