Fixed block parameter destructuring spec
The scenario being tested involves a block receiving more
arguments than there are parameters defined and there is a parameter
for undefined keyword arguments, that is a parameter in the form of **x.
specification expected the ** parameter to be related to the argument
that matched its position in the list.  This was incorrect.  The **
parameter is populated by the last argument in the list when there are
more arguments than parameters defined.