Changed Fiber ensure spec to not be a bug
Since Ruby is what Ruby does, this is not really a bug in Ruby 1.9.x
through Ruby 2.x.  It is the way Ruby behaves.

The behavior specified is that the ensure clause of a fiber does not get
ran.  This is undesirable behavior but it is the behavior.

Relevant Redmine Issues:
  https://bugs.ruby-lang.org/issues/2460
  https://bugs.ruby-lang.org/issues/595