Hash#compare_by_identity behavior changed by frozen string literals
Prior to 2.2 each literal string key was a different object with a
unique identity.  In 2.2 string literal keys were optimized for
Hash#[] and Hash#[]= by using frozen strings as the key and eliminating
duplication which resulted in each unique literal string having only one
identity.

Ruby 2.1 example:
  2.1.5 :001 > x = {}
  => {}
  2.1.5 :002 > x.compare_by_identity
  => {}
  2.1.5 :003 > x["a"] = 1
  => 1
  2.1.5 :004 > x["a"] = 2
  => 2
  2.1.5 :005 > x
  => {"a"=>1, "a"=>2}

Ruby 2.2 example:
  2.2.0 :001 > x = {}
  => {}
  2.2.0 :002 > x.compare_by_identity
  => {}
  2.2.0 :003 > x["a"] = 1
  => 1
  2.2.0 :004 > x["a"] = 2
  => 2
  2.2.0 :005 > x
  => {"a"=>2}

In order to specify this behavior of Hash#assoc the literal string
needed to be forced to duplicate the string key.

Relevant Redmine Issue:
  https://bugs.ruby-lang.org/issues/9382