Errors in Comparable#== are no longer hidden in 2.3
Prior to 2.3, when exceptions occured during #== the exception would be
hidden and the result would be false, indicating the items were not
equal.  This would make it difficult to find related bugs in software.
Beginning with 2.3 these exceptions are being raised so that code will
fail fast and early.

Relevant Redmine Issue:
  https://bugs.ruby-lang.org/issues/7688