Added Bignum specs for unmarshaling 32-bit Bignums on 64-bit platforms
When a 32-bit platform marshals a Bignum, that same number may be
unmarshaled as a Fixnum on 64-bit platforms. We need to ensure that
it still unmarshals correctly on 64-bit platforms.

We ran into the following case:

  val = 1433868532 # a Bignum on 32-bit platforms, Fixnum on 64-bit
  arr = [val, val]
  Marshal.dump(arr)

where `arr` was being dumped by a 32-bit platform, and loaded on
a 64-bit platform, where it threw "dump format error (unlinked)"
because a Bignum is not an ImmediateValue but a Fixnum is.
Since `val` is a Fixnum on 64-bit machines, it was being counted
as an ImmediateValue, and not being added to Marshal's @objects
table when it should have been.