Clean up the entire spec temp dir when finished with mock dirs.
The mock dirs are created with mkdir_p, so they'll create the spec
temp dir if it doesn't already exist. However, it does so after
changing the umask to 0, which affects the permissions on the spec
temp dir. The spec temp dir is normally not recreated if it already
exists, so these atypical permissions persist after any test using
the mock dirs is finished. The safest thing to do is delete the
entire spec temp dir, since the next spec requiring it will cause
it to be recreated, but this time with the corrected umask.