Remove check for unicode_normalize to be required from v2.5 (second try)
The check is no longer needed for Ruby v2.5 and later, because the relevant functionality has been moved to C code in string.c. This check blocks the removal of enc/prelude from the list of files used in the prelude.
(see https://svn.ruby-lang.org/cgi-bin/viewvc.cgi?view=revision&revision=58559 and https://travis-ci.org/ruby/ruby/builds/228602567)
This is a revised version of #433, see there for additional discussion.
This is a revised version of #434, to hopefully get the number of `end`s correct.