Add spec for looped delegation of a block containing a break.
In jruby/jruby#4841 we found that JRuby was incorrectly marking
a delegated block (passed via &block) as having "escaped". We use
this escape bit to know whether non-local flow control like
break has a valid target, and so prematurely marking a delegated
block as escaped caused the second iteration's break to raise an
error. The test here loops twice, which is sufficient to test
that the block's break continues to work on the n+1 iteration.