Time#nsec should always be positive (jruby/jruby#5440)
* Time#nsec should produce a positive result always.

The original code was using dt.getMillis which will produce a
native value for dates before the epoch, causing the final
calculation of nsecs to be both native and of the wrong magnitude.
The correct method to use is dt.getMillisOfSecond, which will
provide a proper positive millis value even before the epoch.

Fixes jruby/jruby#5438.

* Add specs for positive nsec and usec for pre-epoch dates.

See jruby/jruby#5438.