Add tests for exotic rescued error capturing
The idiomatic way to capture exceptions is to assign to a local variable

    rescue => local_variable

However, other kinds of LVALUEs also work

    rescue => $global_variable
    rescue => @@class_variable
    rescue => @instance_variable
    rescue => Constant
    rescue => receiver&.setter_method
    rescue => receiver.setter_method
    rescue => receiver[:key]

Some of the tests involve side effects to the global state. We can
remove the constant and class variable, but the best we can go for the
global variable is to nil it out. This is effectively identical to
removing it though, as nil is the default when accessing an undefined
global.