Add a ruby version guard to the checks of float literals as hash key
This check was introduced in MRI on master branch only in commit
  https://github.com/ruby/ruby/commit/37eb5e74395f148782f7d67b5218fb2e66b113d7

The check works on older ruby versions without the above commit, but only when the system is a 64 bit architecture and is using flonums.
However on 32 bit systems float literals are not warned, which therefore leads to a spec failure.