Add spec for behavior of autoload in Class.new
This behavior is peculiar but can be traced back at least as far
as 1.9.3 (with 1.8.7 exhibiting different behavior). The target
for a bare 'autoload' call in a method in a Class.new block should
be the new class. This differs from constant lookup, which will
skip the new class and use whatever "concrete" class scope is
above it.

This behavior was reported broken in JRuby many years ago and
never completely fixed. The spec provided here happened to work
until a recent fix for autoload scoping broke it (see
jruby/jruby#6749).

There are various problems with this behavior, the most obvious
being that the autoload (used to define a lazy constant) does not
go to the same place that constant lookup happens, and so this
code likely does not do what anyone would want it to do.
Nevertheless, this is current behavior of Ruby.