Add specs for optimized Enumerable#grep{_v} with non-string elements
These are partial ports of the corresponding MRI test [0] for the
`Enumerable#grep` and `#grep_v` optimization. In TruffleRuby it’s not
possible to define a singleton method on a `Regexp` instance because all
such instances are frozen, so we can’t meaningfully translate the part
of that test which checks the behaviour before and after redefining
`#===`, but the rest of it is suitable.

[0] https://github.com/ruby/ruby/commit/d5f0d338c7b5d3d64929b51d29061d369550e8c4#diff-c88c989ca5665e23f5cb9cff6d7d7f6deeb1ffb37d151bf92edc2d15b6acce4bR66