Add specs for UnboundMethod#bind super behavior
UnboundMethod#bind and #bind_call should bind to a virtual include
class below the receiver class when the method is from a module
that does not exist in the receiver's hierarchy. This allows the
rebound method to call `super` to delegate to the receiver's
class and ancestors.

This behavioral change was introduced when `bind_call` was added
but I have not found any tests, specs or documentation about the
behavior change.

See jruby/jruby#7548 and sorbet/sorbet#1188.