Add check for output of stringification of Set
The current tests are far from complete: It checks the recursive case
to see if it does include a `Set: {...}`, and it tests a number of
entries to see if the result is a String.

I choose to use a Set with a single value to ensure ordering is not
relevant (even though Sets in MRI are ordered, I don't think that should
be a requirement). The value is a String, to check if `to_s` and
`inspect` result in the same output.