Only test that sleeps were long enough
This spec was originally added to confirm that sub-millisecond
sleeps actually did sleep; on JRuby and TruffleRuby before fixes
a sleep of 0.0001s would immediately return, because we both used
a sleep function with a minimum resolution of 0.001s. Ideally 100
sleeps of 0.0001s should not exceed 0.03s, but since that's not
the goal of this spec and since it makes the spec flaky under load
or on slower systems, it seems best to remove this check.

The remaining check just confirms that 100x sleep of 0.0001s does
actually sleep for at least 0.01s. Any Ruby failing the spec now
would indicate they are not actually sleeping for 0.0001s and they
need a fix.