Move fiber specs in library/fiber to core
Ruby 3.0 had a fiber extension that added a few extra methods to the Fiber
class, even though these methods were implemented in the core:
* https://github.com/ruby/ruby/blob/ruby_3_0/ext/fiber/fiber.c
* https://github.com/ruby/ruby/blob/ruby_3_0/cont.c#L3122

Ruby 3.1 added these methods to the core, removing the need to require fiber:
* https://github.com/ruby/ruby/blob/ruby_3_1/cont.c#L3262

Since these methods are now part of the core, the specs should move to the core
too.

The resume_spec is a bit of an outlier, this one existed both in core/fiber and
library/fiber, where the library one had an additional dependency on
Fiber#transfer (defined in the ext). These two files have been merged.