Rewrite specs for String#to_c with special float values
The old description said it understands these values, but this is not
true. The only reasons Infinity get parsed as non-zero is because it
starts with an "I", which is interpreted as 1i and that's where the
parsing stops. Any other word that starts with an "I" behaves the same.

The name "special float values" has been borrowed from the
string/to_f_spec.rb