Cleanup of afamily/pfamily specs for Addrinfo#initialize
It looks like there was a lot of copy-paste in this one, often the
description did not match the actual test.
The descriptions have been updated to use the shortest human readable
syntax, so no `Socket::` namespace and no `PF_` prefix to the name. This
matches the afamily specs (or: most of them).

Additionally, a few additional spaces in the descriptions have been
removed.