Fix bigdecimal spec (#1290)
Fix for new feature BigDecimal(float).
Fix assertion of BigMath.log not to expect incorrect random digits to be appended.
Fix assertion of too-strict precision and assertion that expecting precision to be too loose.
Fix divmod, modulo and remainder assertion which was expecting a bug inconsistent with Float as a specification.
Remove coerce expectation to add/sub/mult which was dropped because of a bug, and never implemented correctly in BigDecmal.