Split these specs to test each feature separately
Testing all of these features together does not show useful
information when one of them is not pre-required, and basing the
spec name off the list of expected features means it cannot be
reliably tagged for exclusion.

The change here makes this a separate subprocess run for each
provided library, which improves the spec in a few different ways:

* Descriptions are meaningful and can be individually tagged.
  Previously, any change to the list of features would cause
  the spec description to change and error output did not clearly
  indicate which feature failed.
* Failed requires are represented with the string "error" so that
  failure output is useful. Previous logic would just report that
  the subprocess failed.
* Implementation-specific features do not have to be filtered out.
  Both CRuby and JRuby have features at boot that do not exist in
  other implementations, and may add others in the future. The spec
  should not need updating when that happens.

The require check and the loaded feature check are combined into a
single run for efficiency.